requirements.txt
fastapi
uvicorn[standard]
motor
bcrypt
python-dotenv
email-validator
aiosmtplib